change smoothing to int32

+++ src/shared/main/shared_main.cpp
@@ -30,12 +30,12 @@ Buffer_Report_t queue[BUFFER_SIZE_QUEUE];
 #endif
 #define TURNTABLE_BUFFER_SIZE 16
 #ifdef INPUT_DJ_TURNTABLE_SMOOTHING_LEFT
-int16_t dj_sum_left = 0;
+int32_t dj_sum_left = 0;
 int8_t dj_last_readings_left[TURNTABLE_BUFFER_SIZE];
 int8_t dj_next_left = 0;
 #endif
 #ifdef INPUT_DJ_TURNTABLE_SMOOTHING_RIGHT
-int16_t dj_sum_right = 0;
+int32_t dj_sum_right = 0;
 int8_t dj_last_readings_right[TURNTABLE_BUFFER_SIZE];
 int8_t dj_next_right = 0;
 #endif
